Move `event` module into admin store
Moves the previously named `event` module from the `admin::service`
module to the `admin::store` module, renamed to `events`. This module
holds the implementations of the `AdminServiceEventStore`.
