+++ libsplinter/src/admin/store/events/store/error.rs
