+++ libsplinter/src/admin/store/events/store/mod.rs
@@ -25,7 +25,8 @@ mod error;
 pub mod memory;
 
 pub use self::error::AdminServiceEventStoreError;
-use crate::admin::service::{event::AdminServiceEvent, messages};
+use crate::admin::service::messages;
+use crate::admin::store::events::AdminServiceEvent;
 
 /// Return type of the `AdminServiceEventStore` `list_events_*` methods.
 pub type EventIter = Box<dyn ExactSizeIterator<Item = AdminServiceEvent> + Send>;
