+++ libsplinter/src/store/sqlite.rs
@@ -100,9 +100,9 @@ impl StoreFactory for SqliteStoreFactory {
     #[cfg(feature = "admin-service-event-store-diesel")]
     fn get_admin_service_event_store(
         &self,
-    ) -> Box<dyn crate::admin::service::event::store::AdminServiceEventStore> {
+    ) -> Box<dyn crate::admin::store::events::store::AdminServiceEventStore> {
         Box::new(
-            crate::admin::service::event::store::diesel::DieselAdminServiceEventStore::new(
+            crate::admin::store::events::store::diesel::DieselAdminServiceEventStore::new(
                 self.pool.clone(),
             ),
         )
@@ -114,7 +114,7 @@ impl StoreFactory for SqliteStoreFactory {
     ))]
     fn get_admin_service_event_store(
         &self,
-    ) -> Box<dyn crate::admin::service::event::store::AdminServiceEventStore> {
+    ) -> Box<dyn crate::admin::store::events::store::AdminServiceEventStore> {
         unimplemented!()
     }
 }
